cleaned up systemic hell
