+++ systemic-hell/puzzle-1.c
@@ -1,0 +1,5 @@
+#include "puzzle-1.h"
+
+int mystery_function_1(const char *message) {
+  return 0;
+}
