+++ systemic-hell/puzzle-1.h
@@ -1,0 +1,6 @@
+#ifndef PUZZLE_1_H
+#define PUZZLE_1_H
+
+int mystery_function_1(const char *message);
+
+#endif
