+++ systemic-hell/puzzle-2.c
@@ -1,0 +1,8 @@
+#include "puzzle-2.h"
+
+// I hope you appreciate that I gave you debugger symbols here.
+// I'm evil, but I'm not diabolical. At some level, I think I 
+// want you to succeed.
+int mystery_function_2(const mystery_struct_t *m) {
+  return 0;
+}
