+++ systemic-hell/puzzle-2.h
@@ -1,0 +1,11 @@
+#ifndef PUZZLE_2_H
+#define PUZZLE_2_H
+
+typedef struct mystery_struct_s {
+  struct mystery_struct_s *unknown;
+  char *message;
+} mystery_struct_t;
+
+int mystery_function_2(const mystery_struct_t *m);
+
+#endif
